+++ src/Atata.Cli/GlobalSuppressions.cs
@@ -1,9 +1,7 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-#pragma warning disable S103 // Lines should not be too long
+﻿#pragma warning disable S103 // Lines should not be too long
 
 [assembly: SuppressMessage("Design", "CA1000:Do not declare static members on generic types", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.Cli.ProgramCli`1.InBaseDirectory~`0")]
 [assembly: SuppressMessage("Design", "CA1000:Do not declare static members on generic types", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.Cli.ProgramCli`1.InDirectory(System.String)~`0")]
 [assembly: SuppressMessage("Style", "IDE0032:Use auto property", Justification = "<Pending>", Scope = "member", Target = "~F:Atata.Cli.CliCommand._process")]
-
-#pragma warning restore S103 // Lines should not be too long
+[assembly: SuppressMessage("Minor Code Smell", "S3254:Default parameter values should not be passed as arguments", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.Cli.ProgramCli.ExecuteAsync(System.Threading.CancellationToken)~System.Threading.Tasks.Task{Atata.Cli.CliCommandResult}")]
+[assembly: SuppressMessage("Minor Code Smell", "S3254:Default parameter values should not be passed as arguments", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.Cli.ProgramCli.ExecuteRawAsync(System.Threading.CancellationToken)~System.Threading.Tasks.Task{Atata.Cli.CliCommandResult}")]