+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `public CliCommandResult? Result { get; }` property to `CliCommand`.
 - Add enum `CliCommandKillOnDispose` with values: `None`, `OnlyProcess`, `EntireProcessTree`.
 - Add `public CliCommandKillOnDispose KillOnDispose { get; set; } = CliCommandKillOnDispose.EntireProcessTree` property to `CliCommand`.
+- Add `public Task<CliCommandResult> ExecuteAsync(CancellationToken)` method to `ProgramCli`.
+- Add `public Task<CliCommandResult> ExecuteRawAsync(CancellationToken)` method to `ProgramCli`.
 
 ### Changed